Separate integrations from gobrake versioning
Fixes https://github.com/airbrake/gobrake/issues/189
(Cant install gin)

We shouldn't release new versions of our integrations when we release
gobrake. Meaning, we should decouple integration packages from the core package.

The correct way to install gin, negroni and beego is like the following:

```sh
go get github.com/airbrake/gobrake/gin
go get github.com/airbrake/gobrake/negroni
go get github.com/airbrake/gobrake/beego
```
