+++ beego/go.mod
@@ -1,3 +1,8 @@
 module github.com/airbrake/gobrake/beego
 
 go 1.15
+
+require (
+	github.com/airbrake/gobrake/v5 v5.0.2
+	github.com/astaxie/beego v1.12.3
+)
