+++ examples/negroni/api.go
@@ -5,8 +5,8 @@ import (
 	"log"
 	"net/http"
 
+	ng "github.com/airbrake/gobrake/negroni"
 	"github.com/airbrake/gobrake/v5"
-	ng "github.com/airbrake/gobrake/v5/negroni"
 
 	"github.com/gorilla/mux"
 	"github.com/urfave/negroni"
