+++ gin/go.mod
@@ -1,3 +1,8 @@
-module github.com/airbrake/gobrake/v5/gin
+module github.com/airbrake/gobrake/gin
 
 go 1.15
+
+require (
+	github.com/airbrake/gobrake/v5 v5.0.2
+	github.com/gin-gonic/gin v1.6.3
+)
