+++ negroni/go.mod
@@ -1,7 +1,8 @@
-module github.com/airbrake/gobrake/v5/negroni
+module github.com/airbrake/gobrake/negroni
 
 go 1.15
 
 require (
-	github.com/urfave/negroni v1.0.0 // indirect
+	github.com/airbrake/gobrake/v5 v5.0.2
+	github.com/urfave/negroni v1.0.0
 )
